backend/requirements.txt
django==5.0.6
djangorestframework==3.15.1
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
django-filter==24.2
dj_database_url==2.3.0
six==1.16.0
